move getUpvsIdentity to nases contacts service
