+++ nest-forms-backend/src/nases/services/nases.contacts.service.ts
@@ -4,6 +4,8 @@ import {
   UpvsNaturalPerson,
 } from 'openapi-clients/slovensko-sk'
 
+import ClientsService from '../../clients/clients.service'
+import { ErrorsEnum } from '../../utils/global-enums/errors.enum'
 import ThrowerErrorGuard from '../../utils/guards/thrower-error.guard'
 import { LineLoggerSubservice } from '../../utils/subservices/line-logger.subservice'
 import { NasesErrorsEnum, NasesErrorsResponseEnum } from '../nases.errors.enum'
@@ -58,8 +60,37 @@ export interface CorporateBodyExtractedData {
 export default class NasesContactsService {
   private readonly logger: LineLoggerSubservice
 
-  constructor(private readonly throwerErrorGuard: ThrowerErrorGuard) {
+  constructor(
+    private readonly throwerErrorGuard: ThrowerErrorGuard,
+    private readonly clientsService: ClientsService,
+  ) {
     this.logger = new LineLoggerSubservice(NasesContactsService.name)
+  }
+
+  async getUpvsIdentity(
+    token: string,
+  ): Promise<UpvsNaturalPerson | UpvsCorporateBody | null> {
+    // there is a bug in the container and function `apiUpvsIdentityGet` below, according to 'openapi-clients/slovensko-sk' types
+    // returns information about UpvsNaturalPerson,
+    // in reality it returns information about UpvsCorporateBody as well
+    // after https://github.com/slovensko-digital/slovensko-sk-api/pull/115 is merged, typing can be erased
+    const result = await this.clientsService.slovenskoSkApi
+      .apiUpvsIdentityGet({
+        headers: { Authorization: `Bearer ${token}` },
+      })
+      .then((response) => response.data)
+      .catch((error: unknown) => {
+        this.logger.error(
+          this.throwerErrorGuard.InternalServerErrorException(
+            ErrorsEnum.INTERNAL_SERVER_ERROR,
+            'Failed to get nases identity, verify if this is because of invalid token or a server issue',
+            undefined,
+            error,
+          ),
+        )
+        return null
+      })
+    return result
   }
 
   extractNaturalPersonData(
